requirements.txt
# Core Dependencies
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
jinja2>=3.1.2
click>=8.1.7

# Database (sqlite3 is built into Python)
aiosqlite>=0.19.0

# Web Scraping & APIs
serpapi>=0.1.5
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Content Processing
markdown>=3.5.1
beautifulsoup4>=4.12.2
nltk>=3.8.1
textstat>=0.7.3

# Static Site Generation
pyyaml>=6.0.1
frontmatter>=1.0.0

# Email & Social Media
hubspot-api-client>=7.0.0
tweepy>=4.14.0
linkedin-api>=2.0.0

# Scheduling & Automation
schedule>=1.2.0
celery>=5.3.0
redis>=4.6.0

# Analytics & Monitoring
google-analytics-data>=0.17.0
pandas>=2.1.0
matplotlib>=3.7.2
seaborn>=0.12.2

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
rich>=13.5.2
loguru>=0.7.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
